+++ vscode/webviews/mentions/mentionMenu/MentionMenuItem.tsx
@@ -4,7 +4,6 @@ import {
     FILE_CONTEXT_MENTION_PROVIDER,
     type MentionQuery,
     SYMBOL_CONTEXT_MENTION_PROVIDER,
-    URL_CONTEXT_MENTION_PROVIDER,
     displayLineRange,
     displayPath,
     displayPathBasename,
@@ -40,6 +39,10 @@ import SourcegraphLogo from '../../icons/providers/sourcegraph.svg?react'
 import styles from './MentionMenuItem.module.css'
 
 function getDescription(item: ContextItem, query: MentionQuery): string {
+    if (item.description) {
+        return item.description
+    }
+
     const range = query.range ?? item.range
     const defaultDescription = `${displayPath(item.uri)}:${range ? displayLineRange(range) : ''}`
 
@@ -65,7 +68,8 @@ export const MentionMenuContextItemContent: FunctionComponent<{
     const isOpenCtx = item.type === 'openctx'
     const isFileType = item.type === 'file'
     const isSymbol = item.type === 'symbol'
-    const icon = isSymbol ? (item.kind === 'class' ? 'symbol-structure' : 'symbol-method') : null
+    const icon =
+        item.icon || (isSymbol ? (item.kind === 'class' ? 'symbol-structure' : 'symbol-method') : null)
     const title = item.title ?? (isSymbol ? item.symbolName : displayPathBasename(item.uri))
     const description = getDescription(item, query)
 
@@ -83,9 +87,7 @@ export const MentionMenuContextItemContent: FunctionComponent<{
     return (
         <>
             <div className={styles.row}>
-                {item.type === 'symbol' && icon && (
-                    <i className={`codicon codicon-${icon}`} title={item.kind} />
-                )}
+                {icon && <i className={`codicon codicon-${icon}`} title={isSymbol ? item.kind : ''} />}
                 <span className={clsx(styles.title, warning && styles.titleWithWarning)}>{title}</span>
                 {description && <span className={styles.description}>{description}</span>}
             </div>
@@ -107,7 +109,7 @@ export const MentionMenuProviderItemContent: FunctionComponent<{
     )
 }
 
-export const iconForProvider: Record<
+const iconForProvider: Record<
     string,
     React.ComponentType<{
         size?: string | number
@@ -116,7 +118,6 @@ export const iconForProvider: Record<
 > = {
     [FILE_CONTEXT_MENTION_PROVIDER.id]: FileIcon,
     [SYMBOL_CONTEXT_MENTION_PROVIDER.id]: SquareFunctionIcon,
-    [URL_CONTEXT_MENTION_PROVIDER.id]: LinkIcon,
     // todo(tim): OpenCtx providers should be able to specify an icon string, so
     // we don't have to hardcode these URLs and other people can have their own
     // GitHub provider etc.