+++ vscode/webviews/promptEditor/plugins/atMentions/util.ts
@@ -3,6 +3,8 @@ import { type ContextItem, ContextItemSource } from '@sourcegraph/cody-shared'
 export function contextItemID(item: ContextItem): string {
     return JSON.stringify([
         `${item.type}`,
+        `${item.type === 'repository' ? item.repoID : ''}`,
+        `${item.type === 'tree' ? item.title : ''}`,
         `${item.uri.toString()}`,
         `${item.type === 'symbol' ? item.symbolName : ''}`,
         item.range